More structure updates
Updated the project settings so the build folder and intermediate folder are a bit more obvious as to what they are.
Rename main.cpp to engine.cpp
Added logger, assert, debugbreak, delegate from my sandbox project to this project but with changes to remove dependencies on FMT header and instead use std::format instead
Added Window class, designed to wrap the creation of a Raylib window in a class so we aren't accidentally calling raylib functions in a way we can't easily trace such as creating 2 windows when raylib has limitations for only 1 window.